# Backport DDB connector Credential Provider Fix  for v5.4.0